msvc compilation error fix #6

+++ models/front_wave_1d/inc/front_wave_1d_impl.hpp
@@ -29,7 +29,7 @@ public:
   front_wave_1d_impl(double alpha, double x0 = 1.0, std::size_t seed=0);
   ~front_wave_1d_impl();
 
-  void seed(uint32_t s);
+  void seed(uint64_t s);
   void clear();
   void reinitialize();
